requirements.txt
Pillow==10.1.0
tqdm==4.66.1 